Blinking cursor seems to be working. Further testing needed. Also fixed timers crashing

+++ Eight/src/LuaLibs/Timer.cs
@@ -91,7 +91,7 @@ namespace Eight.LuaLibs {
                 },
                 new() {
                     Type = LuaType.Number,
-                    Value = timerId,
+                    Value = (double) timerId,
                 }
             };
